requirements.txt
django
MySQL-python
numpy
scikit-learn
nltk
python-igraph
termcolor
enum34
djangorestframework
